de-monolith Stage 5: relocate governor_guard + posture_tracker to kirra-core; parko deps the lean crate
Move the two remaining lean shared safety primitives out of the heavy
`kirra-runtime-sdk` into `kirra-core`, both pure-move + re-export:

- `src/governor_guard.rs` (`all_finite`, the #410 non-finite-input convergence
  point) -> `kirra-core`. Byte-identical (zero deps, zero import changes).
- `src/posture_tracker.rs` (`PostureTracker` / `POSTURE_STALENESS_TIMEOUT_MS`, the
  fail-closed fleet-posture state machine) -> `kirra-core`. Verbatim except the
  file-path header comment and the one intra-crate import
  (`crate::verifier::FleetPosture` -> `crate::FleetPosture`); all 315 lines and the
  11 unit tests move unchanged.

Both SDK modules become thin `pub use kirra_core::...::*;` re-exports, so every
existing `kirra_runtime_sdk::governor_guard::*` / `::posture_tracker::*` path keeps
the SAME item — the scalar kernel, the C FFI, the ROS2 adapter's `pub use
kirra_runtime_sdk::posture_tracker`, and parko's diverse governor are all unchanged.

With `FleetPosture` (Stage 1), the kinematics-contract talisman (Stage 3), and now
`governor_guard` + `posture_tracker` all in `kirra-core`, the parko crates get a
direct dependency on the lean foundation: `parko-kirra` and `parko-ros2` now import
the contract / posture / guard from `kirra_core`, and add a direct `kirra-core` dep.

Honest scope: parko KEEPS its `kirra-runtime-sdk` dependency on purpose — `parko-kirra`
(audit_sink, clearance_delivery) and `parko-ros2` (clearance_gate) persist clearances
and the hash-chained audit log through the SQLite `VerifierStore`, which is verifier-tier
code that must not live in lean `kirra-core`. This stage adds the lean dep; it does not
(and should not) remove the heavy one. The shared `kirra-map` crate remains gated on the
adapter de-monolith (the geometry substrate Point/CorridorSource/PerceivedObject still
lives in `kirra-ros2-adapter`), which is the next stage.

Verification: root workspace build + tests green (kirra-core 72 tests incl. the moved
governor_guard + posture_tracker suites); CI-equivalent `cargo clippy --workspace
-- -D warnings` clean; parko workspace touched crates green (parko-core 204,
parko-kirra 148, parko-ros2 168) and clippy clean; no `#[path]` include hazards; both
moves proven verbatim by diff. The only parko failures are the pre-existing parko-onnx
ORT-runtime environment panics (no native ONNX lib in this container), unrelated.
