+++ parko/crates/parko-ros2/src/lib.rs
@@ -38,10 +38,10 @@ pub mod radar_shim;
 pub mod sensor_mapping;
 pub mod tick_pipeline;
 
-// Re-export the kernel's PostureTracker so parko-ros2 consumers can
-// refer to `parko_ros2::PostureTracker` directly — single
-// implementation, shared with kirra-ros2-adapter via the kernel.
-pub use kirra_runtime_sdk::posture_tracker::{PostureTracker, POSTURE_STALENESS_TIMEOUT_MS};
+// Re-export the PostureTracker so parko-ros2 consumers can refer to
+// `parko_ros2::PostureTracker` directly — single implementation, shared with
+// kirra-ros2-adapter via the lean `kirra-core` crate (de-monolith Stage 5).
+pub use kirra_core::posture_tracker::{PostureTracker, POSTURE_STALENESS_TIMEOUT_MS};
 pub use crate::posture_state::{fleet_to_safety, ParkoPostureState};
 
 #[cfg(feature = "ros2")]
